==> Notebooks/01_Creating_Champions_Dict.ipynb <==
import numpy as np
import pandas as pd
import json as js

# 1. Reading in the Data:

champions = pd.read_csv("../Data/Champions_Items_Info/riot_champion.csv")

# 2. Filtering the relevant information:
The most crucial information consists of:
- name (champion name)
- key  (key used in the game data files)
- tags (the type of champion)

Name and key are used to match the chamipons, while tags might be useful later on in order to figure out the roles this champion can fill in.

champions = champions[["key", "name", "tags"]]

champions.head()

champions.dtypes

# 3. Putting the information into a dictionary:
The keys will be changed to strings, so that they dont throw errors later when writing them to the json file. 

champions.astype({"key": "str"})

champ_dict = {}

for i in range(len(champions)):
    champ_dict[int(champions.iloc[i][0])]=[champions.iloc[i][1],eval(champions.iloc[i][2])]

# 3.1 Putting the dictionary into a json and a csv file for later use:

with open('../Data/Champions_Items_Info/champions.json', 'w') as f:
    js.dump(champ_dict, f)

champions.to_csv("../Data/Champions_Items_Info/champions.csv", index=False)
