==> Notebooks/02_Creating_Items_Dict.ipynb <==
import pandas as pd
import numpy as np
import json as js

# 1. Reading the item file:
The relevant columns for describing the items are:
        - item_id
        - name
        - upper_item
        - tag

The *item_id* and the *name* will be used later to match the information on the match data, while the upper item gives the possiblilty to check wether the item is completed or if it can still be upgraded into a more powerful item. The tag shows what kind of item it is, such as providing:
        - health
        - mana
        - regeneration
        - attack damage
        - etc.

items = pd.read_csv("../Data/Champions_Items_Info/riot_item.csv")

items = items[["item_id", "name", "upper_item", "tag"]]

items = items.astype({"item_id":"str"})

# 2. Creating a dicitionary:
the **NaNs** need to be replaced, for **tag** and for **upper item**.

filling_values = {"upper_item":"[]", "tag":""}

items = items.fillna(filling_values)

item_dict = {}

for i in range(len(items)):
    item_dict[items.iloc[i][0]]=[items.iloc[i][1],eval(items.iloc[i][2]),items.iloc[i][3]]

with open('../Data/Champions_Items_Info/items.json', 'w') as f:
    js.dump(item_dict, f)

items.to_csv("../Data/Champions_Items_Info/items.csv", index=False)
