==> Notebooks/03_First_Look_at_Games.ipynb <==
import pandas as pd
import numpy as np
import re

# 1.Reading in the Data:

df_win = pd.read_pickle("../Data/Match_Info/match_winner_data.pickle")

df_lose = pd.read_pickle("../Data/Match_Info/match_loser_data.pickle")

df_match = pd.read_pickle("../Data/Match_Info/match_data.pickle")

# 2. How to access the json dictionary lists:
first check out how to get the summonerName, then figure out how to get the information on the participants.

On the way start filtering only Solo-Q ranked games, which have the **queueId** 420 and the **gameMode** CLASSIC.

df_match=df_match[df_match["gameMode"]=="CLASSIC"]

df_match = df_match[df_match["queueId"]==420]

df_match = df_match.reset_index(drop=True)

### Looking into the participants:
The participants is structured as follows:

1. List of dictionaries
2. Contains 10 players for each game
3. each player gets some further information
4. **stats** and **timeline** are nested dictionaries within the dictionary

In order to find out, which of those to keep, lets take a deeper look into one of the participants and the information conntained:
        

list(df_match["participants"][1][0])

list(df_match["participants"][1][0]["stats"])

list(df_match["participants"][1][0]["timeline"])

Of those entries I am picking a few to keep and drop the rest.

def parse_participant(participant_dict):
    keys_to_keep = ["participantId",'teamId','championId','spell1Id','spell2Id']
    stats_keys = (['win',"item0", "item1","item2","item3","item4","item5","item6", "kills", "deaths", "assists",
                  "doubleKills", "tripleKills", "quadraKills", "pentaKills", "goldEarned","role", "lane"])
    e = {}
    for key, value in participant_dict.items():
        if key in keys_to_keep:
            e[key] = value
        elif type(value) == dict:
            for key2, value2 in value.items():
                if key2 in stats_keys:
                    e[key2] = value2
                    
    return e

parts = df_match["participants"].copy()

for i in range(len(parts)):
    e = [parse_participant(part) for part in parts.loc[i]]
    parts[i] = e

df_match["participants"]=parts

The same is done for the **participantIdentities** column, although I just want to keep the summoner name, in case I want to look at certain summoners and their performance.

def parse_identity(identity):
    e = {}
    keeping_this = ["summonerName"]
    for key,value in identity.items():
        if type(value)==dict:
            for key2,value2 in value.items():
                if key2 in keeping_this:
                    e[key2]=value2
    return e

ident = df_match["participantIdentities"].copy()

for i in range(len(df_match)):
    ident[i] = [parse_identity(part) for part in ident.loc[i]]

df_match["participantIdentities"]=ident

# 3. Dropping unnessesary columns:

df_match = df_match.drop(columns=["status.status_code","status.message", "seasonId", "gameType", "gameMode", "mapId", "platformId"])

df_match.head()

# 4. Changing the version to just show the patch:
Removing the tails of the patch numbers: 10.6.314.4405 --> 10.6

for i in range(len(df_match)):
    df_match["gameVersion"][i] = df_match["gameVersion"][i][:4]

for i in range(len(df_match)):
    df_match["gameVersion"][i]= df_match["gameVersion"][i].strip(".")

df_match["gameVersion"]=df_match["gameVersion"].astype(float)

df_match.to_pickle("../Data/Match_Info/match_data.pickle")
