==> Notebooks/04_Flattening_the_Dataframe.ipynb <==
import pandas as pd
import numpy as np

df_match = pd.read_pickle("../Data/Match_Info/match_data.pickle")

df_match["participantIdentities"]

# Flattening the Dataframe
The runtimes will become very long, thats why the Dataframe will be split into different partitions. This way it is ensured, that the runtime wont crash and the memory keeps up with the work.

The variables in the funciton are:
- **df** is the dataframe used
- **column** is the column to be flattened
- **lower** lower limit
- **upper** upper limit
- **logical** is a logical flag, if set to true, it will take some other columns and append them to the df, if not, it will be used to just flatten the df.


def flattening_df(df,column:str, lower, upper, logical=True):
    flattened_df = pd.DataFrame()
    for i in range(lower,upper):
        if logical == True:
            flattened_df = (flattened_df.append(pd.json_normalize(df.loc[i,column]))
                                       .assign(gameDuration = df.loc[i,"gameDuration"]
                                            ,gameId=df.loc[i,"gameId"]
                                            ,gameVersion=df.loc[i,"gameVersion"]))
        else:
            flattened_df = flattened_df.append(pd.json_normalize(df.loc[i,column]))
    return flattened_df

# 2. Automatization of the Flattening process:
Taking stepsize of 100, that way the runtime stays short respectively to taking a lot of rows at the same time:

participants = pd.DataFrame()

stepsize = 100
for i in range(int(len(df_match)/stepsize)+1):
    lower = i*stepsize
    upper = (i+1)*stepsize
    dataframe = flattening_df(df_match,"participants", lower,upper)
    participants=participants.append(dataframe)
    if lower%1000==0:
        print(lower)

participants.shape

names = pd.DataFrame()

stepsize = 100
for i in range(int(len(df_match)/stepsize)+1):
    lower = i*stepsize
    upper = (i+1)*stepsize
    names=names.append(flattening_df(df_match,"participantIdentities", lower,upper,logical=False))
    if lower%10000==0:
        print(lower)

names.head()

df_merge = participants.assign(summonerName = names['summonerName'])

df_merge.describe()

df_merge.to_pickle("../Data/Match_Info/match_data.pickle")
