==> Notebooks/12_Check_Win_Rates.ipynb <==
import pandas as pd
import numpy as np

match = pd.read_csv("../Data/Match_Data/prepared_dataset.csv")

match.dtypes

champs = pd.read_csv("../Data/Champions_Items_Info/champions.csv")

champs.head()

champ_list = champs["name"].to_list()

df = pd.DataFrame()

# Creating a DataFrame in order to look at following stats for each champion:
        1. total games played
        2. games played blue
        3. games played red
        4. total win rate
        5. win rate blue
        6. win rate red

for item in champ_list:
    print("processing this champion: "+item)
    played_red = []
    played_blue = []
    for i in range(len(match)):
        if item in eval(match["champion_names_blue"].iloc[i]):
            played_blue.append(match.iloc[i,2])
        if item in eval(match["champion_names_red"].iloc[i]):
            played_red.append(match.iloc[i,2])
    tot_games = (len(played_red)+len(played_blue))
    games_red = len(played_red)
    games_blue = len(played_blue)
    if tot_games ==0:
        tot_wr = 0
    else:
        tot_wr = (played_red.count(0)+played_blue.count(1))/tot_games
    if games_blue == 0:
        blue_wr=0
    else:
        blue_wr = played_blue.count(1)/games_blue
    if games_red == 0:
        red_wr = 0
    else:
        red_wr = played_red.count(0)/games_red
    row_df = pd.DataFrame({"champion":[item],"total games":[tot_games], "games blue":[games_blue], "games red":[games_red],\
                           "win rate":[tot_wr], "win rate blue":[blue_wr], "win rate red":[red_wr]})
    
    df = df.append(row_df)

df = df.reset_index(drop=True)

df[df["total games"]>500].sort_values("win rate", ascending=False).head(20).reset_index(drop=True)

df["champion type"]=0
for i in range(len(df)):
    df["champion type"].iloc[i] = champs[champs["name"]==df.iloc[i,0]].iloc[0,2]

df.sort_values("total games", ascending=False)

champs_list = []
for i in range(len(champs)):
    [champs_list.append(tag) for tag in eval(champs["tags"].loc[i]) if tag not in champs_list]

for item in champs_list:
    df[item]=0
    df[item]=0

for i in range(len(df)):
    for item in champs_list:
        if item in eval(df["champion type"].iloc[i]):
            df[item].iloc[i]=1

df.head()

#df.to_csv("../Data/Match_Data/win_rates.csv", index=False)

df.sort_values("total games", ascending=False).head(20)
