==> Notebooks/13_Checking_Win_Rates_of_Champion_Types.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("../Data/Match_Data/prepared_dataset.csv")

champion_types = ['Fighter', 'Tank', 'Mage', 'Support', 'Marksman']

win_rates = pd.DataFrame()

champ_type = []
games_played = []
win_rate = []
number = []
for types in champion_types:
    for i in range(6):
        games_blue = len(df[df[types+"_blue"]==i])
        if games_blue>0:
            percent_blue = df[df[types+"_blue"]==i]["blueWins"].value_counts(1)[1]
            champ_type.append(types)
            games_played.append(games_blue)
            win_rate.append(percent_blue)
            number.append(i)
        else:
            champ_type.append(types)
            games_played.append(0)
            win_rate.append(0)
            number.append(i)
        

rates = pd.DataFrame()

rates["types"]=champ_type

rates["games_no"]=games_played

rates["win_rate"]=win_rate

rates["number"]=number

rates.pivot(index='types', columns='number', values='win_rate').T.to_csv("../Data/Match_Data/wr_types_blue.csv")

rates = pd.DataFrame()
champ_type = []
games_played = []
win_rate = []
number = []
for types in champion_types:
    for i in range(1,6):
        games_red = len(df[df[types+"_red"]==i])
        if games_red>5:
            percent_red = df[df[types+"_red"]==i]["blueWins"].value_counts(1)[0]
            champ_type.append(types)
            games_played.append(games_red)
            win_rate.append(percent_red)
            number.append(i)
        else:
            champ_type.append(types)
            games_played.append(0)
            win_rate.append(0)
            number.append(i)
            
rates["types"]=champ_type
rates["games_no"]=games_played
rates["win_rate"]=win_rate
rates["number"]=number
rates.pivot(index='types', columns='number', values='win_rate').T.to_csv("../Data/Match_Data/wr_types_red.csv")

df["blueWins"].sum()/len(df)

(len(df)-df["blueWins"].sum())/len(df)

len(df)

300/0.0019*100

for i in range(5):
    games_tot = len(df[df["Tank_red"]==i])
    if games_tot ==0:
        games_tot=1
    if len(df[df["Tank_red"]==i]["blueWins"].value_counts())>1:
        wins_blue = df[df["Tank_red"]==i]["blueWins"].value_counts()[0]
    win_rate = wins_blue/games_tot
    print(i)
    print(games_tot)
    print(wins_blue)
    print(win_rate)

for i in range(5):
    games_tot = len(df[df["Tank_blue"]==i])
    if games_tot ==0:
        games_tot=1
    if len(df[df["Tank_blue"]==i]["blueWins"].value_counts())>1:
        wins_blue = df[df["Tank_blue"]==i]["blueWins"].value_counts()[1]
    win_rate = wins_blue/games_tot
    print(i)
    print(games_tot)
    print(wins_blue)
    print(win_rate)

df.head()

print("Blue team:\n")
print("Tank:")
print(df[df["blueWins"]==1]["Tank_blue"].mean())
print("Mage:")
print(df[df["blueWins"]==1]["Mage_blue"].mean())
print("Support:")
print(df[df["blueWins"]==1]["Support_blue"].mean())
print("Attack Damage:")
print(df[df["blueWins"]==1]["Marksman_blue"].mean())
print("Fighter:")
print(df[df["blueWins"]==1]["Fighter_blue"].mean())

print("Red team:\n")
print("Tank:")
print(df[df["blueWins"]==1]["Tank_blue"].std())
print("Mage:")
print(df[df["blueWins"]==1]["Mage_blue"].std())
print("Support:")
print(df[df["blueWins"]==1]["Support_blue"].std())
print("Attack Damage:")
print(df[df["blueWins"]==1]["Marksman_blue"].std())
print("Fighter:")
print(df[df["blueWins"]==1]["Fighter_blue"].std())

df[df["blueWins"]==0][["Tank_blue", "Mage_blue", "Support_blue", "Marksman_blue"]].to_csv("../Data/Match_Data/for_box_plot.csv")
